+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unereleased]
+## [UNRELEASED]
 
 ### Added
 
-- mehtod to reset core system states
+- method to reset core system states
+- app version getter
 
 ## [0.5.1]