+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - aliases for systems without version to simplify interaction with them
+- method `beforeUninstall` which is called before removing plugins
 
 ### Changed